add bloc in news
