+++ lib/pages/news/content_news.dart
@@ -3,6 +3,7 @@ import 'dart:async';
 import 'package:FlutterNews/conection/api.dart';
 import 'package:FlutterNews/pages/news/news_bloc.dart';
 import 'package:FlutterNews/util/bloc_provider.dart';
+import 'package:FlutterNews/widgets/custom_tab.dart';
 import 'package:FlutterNews/widgets/erro_conection.dart';
 import 'package:flutter/material.dart';
 import 'package:FlutterNews/domain/notice/notice.dart';
@@ -28,16 +29,7 @@ class ContentNewsPage extends StatefulWidget{
 
 class _ContentNewsPageState extends State<ContentNewsPage> with TickerProviderStateMixin{
 
-  var current_category = 'geral';
-  List _news = new List();
-  List _categorys = new List();
-  List _category_english = new List();
-  var carregando = false;
-  var repository = new NewsApi();
-  var page = 0;
-  var pages = 1;
-  var category_selected = 0;
-
+  List _categorys = new List<String>();
   NewsBloc bloc;
 
   @override
@@ -49,15 +41,6 @@ class _ContentNewsPageState extends State<ContentNewsPage> with TickerProviderSt
     _categorys.add("Entretenimento");
     _categorys.add("Saúde");
     _categorys.add("Negócios");
-
-    _category_english.add("geral");
-    _category_english.add("sports");
-    _category_english.add("technology");
-    _category_english.add("entertainment");
-    _category_english.add("health");
-    _category_english.add("business");
-
-    //loadCategory(current_category,page);
 
     super.initState();
   }
@@ -169,66 +152,12 @@ class _ContentNewsPageState extends State<ContentNewsPage> with TickerProviderSt
 
   Widget _getListCategory(){
 
-    ListView listCategory = new ListView.builder(
-        itemCount: _categorys.length,
-        scrollDirection: Axis.horizontal,
-        itemBuilder: (context, index){
-          return _buildCategoryItem(index);
-        }
-    );
-
-    return new Container(
-      height: 50.0,
-      child: listCategory,
-      color: Colors.grey[200].withAlpha(200),
-    );
-
-  }
-
-  Widget _buildCategoryItem(index){
-
-    return new InkWell(
-      onTap: (){
-        onTabCategory(index);
-        print("click");
+    return CustomTab(
+      itens: _categorys,
+      tabSelected: (index){
+        bloc.setCategoryPosition(index);
       },
-      child: new Row(
-        mainAxisSize: MainAxisSize.min,
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: <Widget>[
-          new Container(
-            margin: new EdgeInsets.only(left: 10.0),
-            child: new Material(
-              elevation: 2.0,
-              color: category_selected == index ? Colors.blue[800]:Colors.blue[500],
-              borderRadius: const BorderRadius.all(const Radius.circular(25.0)),
-              child:  new Container(
-                padding: new EdgeInsets.only(left: 12.0,top: 7.0,bottom: 7.0,right: 12.0),
-                child: new Text(_categorys[index],
-                  style: new TextStyle(
-                      color: Colors.white),
-                ),
-              ),
-            ),
-          )
-        ],
-      ),
     );
-
-  }
-
-  onTabCategory(index){
-
-    if(!carregando) {
-
-      setState(() {
-        category_selected = index;
-        page = 0;
-      });
-
-      loadCategory(_category_english[index], page);
-
-    }
 
   }
 
@@ -236,63 +165,6 @@ class _ContentNewsPageState extends State<ContentNewsPage> with TickerProviderSt
 
     bloc.load(false);
 
-    return null;
-  }
-
-  loadCategory(category,page) async{
-
-    if(page < pages-1 || page == 0) {
-
-//      setState((){
-//
-//        current_category = category;
-//
-//        if(page == 0) {
-//          _news.clear();
-//        }
-//
-//        carregando = true;
-//
-//      });
-
-      Map result = await repository.loadNews(category, page.toString());
-
-      if(result != null) {
-
-        widget.errorConection = false;
-
-        setState(() {
-          pages = result['data']['pages'];
-          this.page = page + 1;
-          result['data']['news'].forEach((item) {
-            var notice = new Notice(
-                item['url_img'] == null ? '' : item['url_img'],
-                item['tittle'] == null ? '' : item['tittle'],
-                item['date'] == null ? '' : item['date'],
-                item['description'] == null ? '' : item['description'],
-                item['category'] == null ? '' : item['category'],
-                item['link'] == null ? '' : item['link'],
-                item['origin'] == null ? '' : item['origin'],
-            );
-            _news.add(notice);
-
-          });
-
-          carregando = false;
-        }
-
-        );
-      }else{
-
-        widget.errorConection = true;
-
-        setState((){
-          carregando = false;
-        });
-
-      }
-
-    }
   }
 
   @override
